refactor: extract useHomePageState hook, ShelterPopup, and RefreshErrorToast
- Extract state management, sorting, filtering, and routing logic
  from App.tsx into useHomePageState custom hook
- Extract map popup UI from Map.tsx into ShelterPopup component
- Extract refresh error toast from App.tsx into RefreshErrorToast component
- App.tsx: 333 → 185 lines, Map.tsx: 395 → 310 lines
